backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-encrypted-model-fields>=0.6
django-filter>=24.2
django-celery-beat>=2.6
django-celery-results>=2.5
celery[redis]>=5.3
channels>=4.1
channels-redis>=4.2
redis>=5.0
psycopg2-binary>=2.9
paramiko>=3.4
pywinrm>=0.4
deepdiff>=6.7
jsonschema>=4.21
gunicorn>=21.2
python-decouple>=3.8
debugpy>=1.8
cryptography>=42.0

# ── Authentication ────────────────────────────────────────────────────────────
# LDAP
django-auth-ldap>=4.8
python-ldap>=3.4

# SAML 2.0
djangosaml2>=1.9


